% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My function to create a barplot of a binomial distribution}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations for the sample to run}

\item{n}{The number of samples to take}

\item{p}{The probability of success}
}
\description{
Creates a barplot showing the number of sucesses after a given number of iterations
}
