% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 to create a bootstrap interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{function identifier to apply}

\item{alpha}{confidence interval}

\item{cx}{character expansion factor}
}
\description{
Creates a bootstrap interval with a
a given sample, function and confidence
interval
}
