% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My clt creates a histogram of sums}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{a}{minimum value of uniform}

\item{b}{maximum value of uniform}
}
\description{
Creates a histogram of sums from the random variable
uniform distribution for a given a and b,
with sample size n over iter iterations
}
