% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg function creates a maximum likelihood estimate}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{function name to use}

\item{theta}{sequence of theta to create the estimate for}
}
\description{
creates a maximum likelihood estimate to find graphical max
}
