% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My normal curve creates a normal distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{a}{y value to calculate P(Y <= a)}
}
\description{
Creates a normal distribution with a given mean and standard
deviation. Then, it shades the area of the lower tail below a
given y value, as well as outputing the probability of that area
}
